refactor(MapDuck): rename
