+++ src/base/static/components/molecules/left-sidebar-section.tsx
@@ -21,7 +21,7 @@ import {
   LeftSidebarSection,
   LeftSidebarOption,
 } from "../../state/ducks/left-sidebar";
-import { MapSourcesLoadStatus } from "../../state/ducks/map-config";
+import { MapSourcesLoadStatus } from "../../state/ducks/map";
 
 const MapLayerSelectorContainer = styled("div")({
   display: "flex",
