+++ src/base/static/components/molecules/map-filter-slider.tsx
@@ -3,7 +3,7 @@ import * as React from "react";
 import { css, jsx } from "@emotion/core";
 
 import { LayerGroup } from "../../state/ducks/map-style";
-import { FilterSliderConfig } from "../../state/ducks/map-config";
+import { FilterSliderConfig } from "../../state/ducks/map";
 import { RangeInput } from "../atoms/input";
 import { RegularText } from "../atoms/typography";
