+++ src/base/static/components/molecules/map-radio-menu.tsx
@@ -10,7 +10,7 @@ import {
   LayerGroups,
   Layer,
 } from "../../state/ducks/map-style";
-import { RadioMenuConfig, RadioMenuOption } from "../../state/ducks/map-config";
+import { RadioMenuConfig, RadioMenuOption } from "../../state/ducks/map";
 import FormLabel from "@material-ui/core/FormLabel";
 import FormControlLabel from "@material-ui/core/FormControlLabel";
 import Radio from "@material-ui/core/Radio";
