+++ src/base/static/components/organisms/offline-download-menu.js
@@ -11,7 +11,7 @@ import {
   ModalFooter,
   modalStyles,
 } from "../atoms/layout";
-import { offlineConfigPropType } from "../../state/ducks/map-config";
+import { offlineConfigPropType } from "../../state/ducks/map";
 import { mapSourcesPropType, mapSourcesSelector } from "../../state/ducks/map-style";
 
 import Modal from "react-modal";
