+++ src/base/static/components/templates/map.tsx
@@ -52,7 +52,7 @@ import {
   mapConfigSelector,
   mapConfigPropType,
   MapSourcesLoadStatus,
-} from "../../state/ducks/map-config";
+} from "../../state/ducks/map";
 import {
   createFeaturesInGeoJSONSource,
   mapSourcesSelector,
