+++ src/base/static/state/ducks/map.d.ts
