+++ src/base/static/state/ducks/map.js
